feat(app): change sms invitaton implementation
