+++ app/models/invitation.rb
@@ -4,11 +4,7 @@ class Invitation < ApplicationRecord
 
   enum format: { sms: 0, email: 1, link_only: 2 }
 
-  def send_invite
-    SendInvitation.call(invitation: self)
-  end
-
-  def as_json(_opts)
+  def as_json(_opts = {})
     super.merge(sent_at: sent_at&.to_date&.strftime("%m/%d/%Y"))
   end
 end
