+++ app/models/rdv_solidarites_user.rb
@@ -1,6 +1,6 @@
 class RdvSolidaritesUser
   USER_ATTRIBUTES = [
-    :id, :first_name, :last_name, :birth_date, :email, :phone_number,
+    :id, :first_name, :last_name, :birth_date, :email, :phone_number, :phone_number_formatted,
     :birth_name, :address, :affiliation_number, :created_at, :invited_at
   ].freeze
