--- app/services/send_sms_invitation.rb
@@ -1,35 +1,0 @@
-class SendSmsInvitation < BaseService
-  def initialize(invitation:)
-    @invitation = invitation
-  end
-
-  def call
-    send_sms
-  end
-
-  private
-
-  def send_sms
-    Rails.logger.info(content)
-    return if Rails.env.development?
-
-    SendTransactionalSms.call(applicant.phone_number, content)
-  end
-
-  def content
-    "Bonjour,\n Vous êtes allocataire du RSA. Vous devez bénéficier d'un accompagnement obligatoire dans " \
-    "le cadre de vos démarches d'insertion. Le département #{department.name.capitalize} vous invite à prendre " \
-    "rendez-vous auprès d'un référent afin d'échanger sur votre situation.\n" \
-    "Vous devez prendre rendez-vous en ligne à l'adresse suivante: #{@invitation.link}\n" \
-    "En cas d'absence, une sanction pourra être prononcée. Pour tout problème, contactez " \
-    "le secrétariat au #{department.phone_number}."
-  end
-
-  def department
-    @invitation.department
-  end
-
-  def applicant
-    @invitation.applicant
-  end
-end
