+++ config/initializers/send_in_blue.rb
@@ -1,0 +1,3 @@
+SibApiV3Sdk.configure do |config|
+  config.api_key['api-key'] = ENV['SENDINBLUE_API_V3_KEY']
+end
